feat(DQN): Move PPO models to a common directory
